% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addDen}
\alias{addDen}
\title{Add Normal Density Curve to persp Plot in z Dimension}
\usage{
addDen(x, mu, sigma, ylim, persp, n = 50, perc = 95, incol = "light blue",
  outcol = "blue", density = 40)
}
\arguments{
\item{x}{x coordinate.}

\item{mu}{mean of normal ditribution.}

\item{sigma}{standard deviation of normal distribution.}

\item{ylim}{limits over which to draw density}

\item{persp}{output of \code{\link{persp}}.}

\item{n}{number of points at which to evaluate density.}

\item{perc}{middle percentage to colour by \code{incol}.}

\item{incol}{colour for middle \code{perc} percent of density.}

\item{outcol}{colour for density outside middle \code{perc} percent.}

\item{density}{density of hashing.}
}
\description{
Add normal density curve to persp plot in z dimension.
}
\author{
Heather Turner
}
\seealso{
\code{\link{polygon}}
}

