% Generated by roxygen2 (4.0.2): do not edit by hand
\name{base3D}
\alias{base3D}
\title{Add Box to Base of 3D Plot}
\usage{
base3D(xlim, ylim, zlim, P)
}
\arguments{
\item{xlim}{x limits.}

\item{ylim}{y limits.}

\item{zlim}{z limits.}

\item{P}{output of \code{\link{persp}}.}
}
\description{
Add box to base of 3D plot.
}
\author{
Heather Turner
}

