% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataViz.R
\name{dataViewer}
\alias{dataViewer}
\alias{dataViewerUI}
\alias{dataViewerServer}
\title{shiny UI to view data}
\usage{
dataViewerUI(id)

dataViewerServer(input, output, session, data = reactive(NULL),
  part = "all", style = "minimal")
}
\arguments{
\item{id}{namespace identifier for the module}

\item{input, output, session}{mandatory arguments for modules to be valid. These
should not to be defined as they will be handled by shiny.}

\item{data}{the input data.frame to be displayed. It must be a reactive
value.}

\item{part}{one of "all" to show all table, c("header", "footer") to show first and last rows, "header"
to show only first rows, "footer" to show only last rows.}

\item{style}{define how to display the DataTable ("minimal" or "default")}
}
\description{
A module to view data in a DataTable output. An option is available
to display only first or last rows of the dataset.
}
\examples{
library(shinytools)
library(shiny)

# example with dataViewerUI and dataViewerServer ----
if (interactive()) {
  options(device.ask.default = FALSE)

  ui <- fluidPage(
    fluidRow(column(width = 12, dataViewerUI(id = "demo") ) )
  )

  server <- function(input, output, session) {
    callModule(module = dataViewerServer, id = "demo",
               data = reactive(iris), part = c("header", "footer"))
  }
  print(shinyApp(ui, server))
}

}
