% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dataViz.R
\name{dataVizServer}
\alias{dataVizServer}
\title{Module dataViz Server part}
\usage{
dataVizServer(input, output, session, data = reactive(NULL),
  part = "all", style = "minimal")
}
\arguments{
\item{input}{Not a real parameter, not to be set manually.}

\item{output}{Not a real parameter, not to be set manually.}

\item{session}{Not a real parameter, not to be set manually.}

\item{data}{Reactive data.frame}

\item{part}{"All" to show all table, "header" & "footer" to show head and/or tail}

\item{style}{Style that defines how to display DT table (only minimal & default available for the moment)}
}
\description{
Server part of module dataViz
}
