% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability.R
\name{default_disabled}
\alias{default_disabled}
\title{add disabled tag to an HTML element.}
\usage{
default_disabled(x)
}
\arguments{
\item{x}{html element}
}
\description{
The function is to be used to disable an element when created.
}
\examples{
library(shinytools)

if (interactive()) {
  options(device.ask.default = FALSE)

  ui <- fluidPage(
    titlePanel("Hello dghiesse!"),
    load_jstools(),
    sidebarLayout(
      sidebarPanel(
        default_disabled(actionButton("dosomething", "don't click"))
      ),
      mainPanel(
        plotOutput("distPlot")
      )
    )
  )

  server <- function(input, output) {
    output$distPlot <- renderPlot({
      hist(rnorm(50))
    })
  }

  print(shinyApp(ui, server))
}

}
