% Generated by roxygen2 (4.0.2): do not edit by hand
\name{densStripPois}
\alias{densStripPois}
\title{Add Poisson Density Strip}
\usage{
densStripPois(fit, perc = 95, nx = 128, ny = 128, persp = NULL,
  xlim = NULL, ylim = NULL, col = "red")
}
\arguments{
\item{fit}{list with components \code{a} and \code{b}, the intercept
and slope of the fit.}

\item{perc}{middle percentage of density shown.}

\item{nx}{number of x values to grid over.}

\item{ny}{number of y values to grid over.}

\item{persp}{the output from \code{persp} to transform to 3D.}

\item{xlim}{x limits to clip to in 3D case.}

\item{ylim}{y limits to clip to in 3D case.}

\item{col}{colour of the density image.}
}
\description{
Adds a Poisson density strip around E(y) to existing 2D or 3D plot of y vs x,
given the slope and intercept from a Poisson regression.
}
\details{
Density computed over \code{nx} by \code{ny} grid of points and displayed as
an image plot.
}
\author{
Heather Turner
}

