% Generated by roxygen2 (4.0.2): do not edit by hand
\name{denStrip}
\alias{denStrip}
\title{Add Normal Density Strip}
\usage{
denStrip(x, yhat, sigma, npoly = 50, perc = 95, persp = NULL,
  xlim = NULL, ylim = NULL, col = "red")
}
\arguments{
\item{x}{x values (minimum 3 to enable clipping).}

\item{yhat}{fitted values, i.e. E(y).}

\item{sigma}{standard deviation of density.}

\item{npoly}{number of polygons.}

\item{perc}{middle percentage of density shown.}

\item{persp}{the output from \code{persp} to transform to 3D.}

\item{xlim}{x limits to clip to in 3D case.}

\item{ylim}{y limits to clip to in 3D case.}

\item{col}{colour of the density polygons.}
}
\description{
Adds a normal density strip around E(y) to existing 2D or 3D plot of y vs x.
}
\details{
Density built up by layering \code{npoly} polygons.
}
\author{
Heather Turner
}

