% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{exportDataUI}
\alias{exportDataUI}
\alias{exportDataServer}
\title{exportData}
\usage{
exportDataUI(id)

exportDataServer(input, output, session, display = c("panel", "button",
  "modal"), ui_label = ifelse(display == "panel", "Export", "Download"),
  modal_title = h1("Download dataset"), types = c("excel", "csv",
  "sas"), datas = NULL)
}
\arguments{
\item{id}{namespace identifier for the module}

\item{input, output, session}{mandatory arguments for modules to be valid. These
should not to be defined as they will be handled by shiny.}

\item{display}{One of c("panel", "button", "modal")

panel : Create a panel with export options

button : Create an actionButton that trigger a modal

modal : When function exportDataServer is called then display a modal
to be used directly into observeEvent handler}

\item{ui_label}{Title of actionButton or panel}

\item{modal_title}{Title of the modal}

\item{types}{Potential export file type (available : excel, csv & sas)}

\item{datas}{Reactive data.frame}
}
\description{
exportData
}
