% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{filled}
\alias{filled}
\title{Check if input filled}
\usage{
filled(x)
}
\arguments{
\item{x}{UI input or reactiveValues}
}
\description{
Unlike req, the filled function doesn't exit block if FALSE
}
\examples{
library(shinytools)
library(shiny)

if (interactive()) {
  options(device.ask.default = FALSE)

ui <- fluidPage(
  fluidRow(
    column(2,
      textInput("TI", label = "", placeholder = "Write something"),
      textOutput("TO_ti")
    ),
    column(2,
      selectInput("SI", label = "Choose one or multiple letters", 
        choices = letters, selected = "", multiple = TRUE),
      textOutput("TO_si")
    ),
    column(2,
      actionButton("AB2", label = "Create reactiveValues"),
      textOutput("TO_reactiveValues")
    )
  )
)

server <- function(input, output) {
  
  rv <- reactiveValues()
  
  output$TO_ti <- renderText({
    if (filled(input$TI)) {
      paste("textInput filled")
    } else {
      paste("textInput empty")
    }
  })
  output$TO_si <- renderText({
    if (filled(input$SI)) {
      paste("selectInput filled")
    } else {
      paste("selectInput empty")
    }
  })
  
  output$TO_reactiveValues <- renderText({
    if (filled(rv$ex)) {
      paste("reactiveValues rv$ex filled")
    } else {
      paste("reactiveValues rv$ex empty")
    }
  })
  
  observeEvent(input$AB2, {
    rv$ex <- input$AB2
  })
}

  print(shinyApp(ui, server))
}
}
