% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDf.R
\name{filterDfServer}
\alias{filterDfServer}
\title{filterDf}
\usage{
filterDfServer(input, output, session, x = reactive(NULL),
  default_show = TRUE, show_all = TRUE, return_datas = FALSE)
}
\description{
filterDf
}
\examples{
library(shinytools)
library(shiny)
library(lubridate)
library(rlang)
if (interactive()) {
  options(device.ask.default = FALSE)


  ui <- fluidPage(
    filterDfUI(id = "demo")
  )

  server <- function(input, output) {
      
      res <- callModule(module = filterDfServer, id = "demo", x = reactive(iris))
      
      observe({
          tmp <- reactiveValuesToList(res)
          if (tmp$filtered) {
            print(tmp$expr)
          }
      })
  }
  print(shinyApp(ui, server))
}
}
