% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCaptures.R
\name{GetCaptures}
\alias{GetCaptures}
\title{GetCaptures}
\usage{
GetCaptures(s, pat)
}
\arguments{
\item{s}{chr string}

\item{pat}{chr string containing a regular expression with at least 1 capture '(', ')'}
}
\value{
Num matrix indexing substrings of s. Each columns represents a substring.
         Row 1 is the label, Row 2 is the start, row 3 the stop letter of each substring.
         Substrings which are part of a capture are labelled with increasing numbers as they appear in s beginning with 1.
         Substrings which are not part of a capture are labelled with 0.
         If nothing was captured in chr s, NULL is returned.
}
\description{
Split a chr string by capture groups according to a regular expression provided.
}
\details{
The string is returned as a matrix of start and stop position of substrings.
These substrings are labelled by which capture group they belong to.
Substrings outside a capture are labelled as 0.

Nested capturing groups do not work.
}
\examples{
s <- ">ENSG00000139083_8_61387.34117"
pat <- "^.*(SG.+?)1.*_(.+?)_.*$"

m <- GetCaptures(s, pat)
apply(m, 2, function(x) substr(s, x[2], x[3]))

}
\seealso{
Other CapturePattern module functions: \code{\link{CapturePatternUI}},
  \code{\link{CapturePattern}}
}

