% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLim}
\alias{getLim}
\title{Get Axis Limits}
\usage{
getLim(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a vector of axis limits.
}
\description{
Get regular axis limits for given variable, i.e. with par("xaxs") or
par("yaxs") equal to "r".
}
\author{
Heather Turner
}

