% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{html_class}
\alias{html_class}
\alias{html_unclass}
\alias{html_addclass}
\title{Add or remove a class from an HTML element}
\usage{
html_class(id, classname, add = TRUE)

html_unclass(id, classname)

html_addclass(id, classname)
}
\arguments{
\item{id}{shiny input id}

\item{classname}{class name}

\item{add}{if TRUE, class will be added, if FALSE class will be deleted}
}
\description{
\code{html_class} will add or remove a class from an HTML element.
\code{html_addclass} and \code{html_unclass} are shortcut functions.
}
\examples{
library(shinytools)
classes <- c("primary", "secondary", "success",
             "danger", "warning", "info")

if (interactive()) {
  options(device.ask.default = FALSE)

  ui <- fluidPage(
    titlePanel("Hello dghiesse!"),
    load_jstools(),
    sidebarLayout(
      sidebarPanel(
        selectInput("class", "Variable:", classes)
      ),
      mainPanel(
        actionButton("anybutton", "watch this button",
                     class = "btn-info")
      )
    )
  )

  server <- function(input, output) {
    observeEvent(input$class, {
      for(i in classes){
        html_unclass("anybutton", paste0("btn-", i))
      }
      html_addclass("anybutton", paste0("btn-", input$class))
    })
  }

  print(shinyApp(ui, server))
}
}
\seealso{
Other javascript functions: \code{\link{ability}},
  \code{\link{activate}}, \code{\link{click_event}},
  \code{\link{html_toogle}}
}
\concept{javascript functions}
