% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toogle.R
\name{html_toogle}
\alias{html_toogle}
\alias{html_set_visible}
\alias{html_set_hidden}
\title{Display or hide an element}
\usage{
html_toogle(id)

html_set_visible(id)

html_set_hidden(id)
}
\arguments{
\item{id}{shiny input id}
}
\description{
\code{html_toogle} will display an element if hidden or hide it
if visible.
\code{html_set_visible} will make an element visible.
\code{html_set_hidden} will hide an element.
}
\examples{
library(shinytools)

if (interactive()) {
  options(device.ask.default = FALSE)

  ui <- fluidPage(
    titlePanel("Hello dghiesse!"),
    load_jstools(),
    sidebarLayout(
      sidebarPanel(
        sliderInput( "obs",
          "Number of observations:",
          min = 0, max = 1000, value = 500),
        actionButton("toogle", "toogle"),
        actionButton("set_visible", "show"),
        actionButton("set_hidden", "hide")
      ),
      mainPanel(
        plotOutput("distPlot")
      )
    )
  )

  server <- function(input, output) {
    output$distPlot <- renderPlot({
      hist(rnorm(input$obs))
    })
    observeEvent(input$toogle, {
      html_toogle("distPlot")
    })
    observeEvent(input$set_visible, {
      html_set_visible("distPlot")
    })
    observeEvent(input$set_hidden, {
      html_set_hidden("distPlot")
    })
  }

  print(shinyApp(ui, server))
}
}
\seealso{
Other javascript functions: \code{\link{ability}},
  \code{\link{activate}}, \code{\link{click_event}},
  \code{\link{html_class}}
}
\concept{javascript functions}
