% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importDataUI}
\alias{importDataUI}
\title{Factor to character}
\usage{
importDataUI(id)
}
\arguments{
\item{id}{Module call id}

\item{label}{Label of actionLink}
}
\description{
UI part of module import
}
