% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Init}
\alias{Init}
\title{Init}
\usage{
Init(args = commandArgs(TRUE), log = paste("Starting at", Sys.time()))
}
\arguments{
\item{args}{chr of path to *.status file (commandArgs(TRUE)).
If the R script was started using RProcess module, this is the first trailing argument of Rscript (default).}

\item{log}{chr arr used as initial log entry for this process. Each element is a line.}
}
\value{
input object as defined in RProcess module
}
\description{
If a R script was started by shiny using the RProcess module, this should be the first command of the script.
This function reads the *.status file to retrieve information about the process and load the input.
As a side effect an object 'info' of class 'ProcessInfo' is written into the global envir.
This is a list which contains information about the current process. It is used by other functions of this module.
}
\details{
A *.status file is used for communication between the shiny session and the R batch process (started with Rscript).
Information provided by RProcess module is handed over to the R batch process via this file.
Conversely the R batch process writes on this file to hand over status updates to the shiny session.
By using the functions \code{\link{Init}}, \code{\link{Up}}, \code{\link{Log}}, \code{\link{Try}},
\code{\link{Fin}} it is ensured that the communication bewteen shiny session and R batch script work and
that no files are falsly overwritten.

If a log file exists, it is appended by the lines defined with log. Each element of log is a line.
Information about path of the log file, the session id and the process id is taken from the ProcessInfo object.
If a session id exists, every line is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the *.status file.
If they do not match, quit() is run.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Abort}},
  \code{\link{Fin}}, \code{\link{Log}},
  \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{ReadInfo}},
  \code{\link{Start}}, \code{\link{Try}}, \code{\link{Up}}
}

