% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_jstools.R
\name{load_jstools}
\alias{load_jstools}
\title{load shinytools JavaScript functions}
\usage{
load_jstools()
}
\description{
load shinytools JavaScript functions in a shiny application.
Operation has to be made in the UI part.
}
\examples{
\dontrun{
library(shinytools)

if (interactive()) {
  options(device.ask.default = FALSE)

  # Define UI
  ui <- fluidPage(
    titlePanel("Hello dghiesse!"),
    load_jstools(),
    sidebarLayout(
      sidebarPanel(
        sliderInput( "obs",
          "Number of observations:",
          min = 0, max = 1000, value = 500)
      ),
      mainPanel(plotOutput("distPlot"))
    )
  )

  server <- function(input, output) {
    output$distPlot <- renderPlot({
      hist(rnorm(input$obs))
    })
  }

  print(shinyApp(ui, server))
}
}

}
