% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Log}
\alias{Log}
\title{Log}
\usage{
Log(msg, info = NULL)
}
\arguments{
\item{msg}{chr arr with log messages. Each element is a line.}

\item{info}{ProcessInfo object or NULL (NULL).
This is a list of relevant information about the process.
It can be created with \code{\link{RProcessInit}}.
If NULL the global environment is searched for an object of class ProcessInfo.}
}
\value{
TRUE if a log was written, FALSE if not
}
\description{
This is a convenience function for writing a log message.
This should be used in a R script started by shiny with the RProcess module.
}
\details{
If a log file exists, it is appended by the lines defined in with msg. Each element of msg is a line.
Information about path of the log file, the session id and the process id is taken from the ProcessInfo object (see arguments).
If a session id exists, every line is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the *.status file.
If they do not match, quit() is run.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.

By using the functions \code{\link{Init}}, \code{\link{Up}}, \code{\link{Log}}, \code{\link{Try}},
\code{\link{Fin}} it is ensured that the communication bewteen shiny session and R batch script work and
that no files are falsly overwritten.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Abort}},
  \code{\link{Fin}}, \code{\link{Init}},
  \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{ReadInfo}},
  \code{\link{Start}}, \code{\link{Try}}, \code{\link{Up}}
}

