% Generated by roxygen2 (4.0.2): do not edit by hand
\name{perspAxis}
\alias{perspAxis}
\title{Add Axes to persp Plot}
\usage{
perspAxis(axis = 1:3, persp, xlim, ylim, zlim)
}
\arguments{
\item{axis}{vector specifying one or more axes to add.}

\item{persp}{the output from \code{persp} to transform to 3D.}

\item{xlim}{x limits.}

\item{ylim}{y limits.}

\item{zlim}{z limits.}
}
\description{
Add axes to persp plot.
}
\author{
Heather Turner
}

