% Generated by roxygen2 (4.0.2): do not edit by hand
\name{perspLab}
\alias{perspLab}
\title{Add Axis Labels to persp Plot}
\usage{
perspLab(persp, xlim = NULL, ylim = NULL, zlim = NULL, xlab = NULL,
  ylab = NULL, zlab = NULL, ...)
}
\arguments{
\item{persp}{the output from \code{persp} to transform to 3D.}

\item{xlim}{x limits.}

\item{ylim}{y limits.}

\item{zlim}{z limits.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{zlab}{label for z axis.}

\item{...}{further arguments passed to \code{text}.}
}
\description{
Add axis labels to persp plot.
}
\author{
Heather Turner
}
\references{
Based on \url{http://rwiki.sciviews.org/doku.php?id=tips:graphics-3d:zlabspace}.
}

