% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactiveTextInput.R
\name{ReactiveTextInputUI}
\alias{ReactiveTextInputUI}
\title{ReactiveTextInputUI}
\usage{
ReactiveTextInputUI(id, title)
}
\arguments{
\item{id}{chr id of this object for shiny session}

\item{title}{chr or NULL for h3 title above text input fields}
}
\value{
chr HTML for creating ui elements.
}
\description{
This is a reactive series of text input fields with a functionality for validating the provided text inputs.
The number of text input fields in reactive.
}
\examples{
library(shinyTools)

# some function as example
check <- function(text, add){ if(any(grepl(add$pat, text))) return(paste("Don't use letter", add$pat, "in any entry."))}

# little app with module
ui <- fluidPage(sidebarLayout(
  sidebarPanel(h2("ReactiveTextInputUI"),
               numericInput("nGroups", "n Groups", 2, -2, 10),
               ReactiveTextInputUI("id1", "Groups")
  ),
  mainPanel(h2("Output of ReactiveTextInput"), verbatimTextOutput("display"))
))


server <-function(input, output, session) {
  display <- callModule(ReactiveTextInput, "id1", n = reactive(input$nGroups), prefix = "Group",
                        values = c("Untreated", "Treated"), checkFun = "check", addArgs = list(pat = "X"))
  output$display <- renderPrint(display())
}

shinyApp(ui, server)

}
\seealso{
Other ReactiveTextInput module functions: \code{\link{ReactiveTextInput}}
}

