% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{ReadInfo}
\alias{ReadInfo}
\title{ReadInfo}
\usage{
ReadInfo(statusFile)
}
\arguments{
\item{statusFile}{chr path to relevant *.status file}
}
\value{
list of vectors of class ProcessInfo.
        This list contains information about the current process and is used by other Rscript communication functions.
}
\description{
ReadInfo
}
\seealso{
These Rscript communication functions are used in R scripts started by a RProcess module such as \code{\link{RProcess}}.
         For some examples on how they work together see the vignette \emph{RProcess Module Functions}.

Other Rscript communication functions: \code{\link{Rscript_Abort}},
  \code{\link{Rscript_Fin}}, \code{\link{Rscript_Init}},
  \code{\link{Rscript_Log}}, \code{\link{Rscript_Start}},
  \code{\link{Rscript_Try}}, \code{\link{Rscript_Up}}
}
\keyword{";".}
\keyword{*.status}
\keyword{Each}
\keyword{ProcessInfo.}
\keyword{Read}
\keyword{Rscript}
\keyword{The}
\keyword{This}
\keyword{\code{*.status}}
\keyword{a}
\keyword{and}
\keyword{are}
\keyword{as}
\keyword{by}
\keyword{class}
\keyword{communicate}
\keyword{communication}
\keyword{content}
\keyword{different}
\keyword{element}
\keyword{entries}
\keyword{entry}
\keyword{file}
\keyword{first}
\keyword{following}
\keyword{function}
\keyword{functions.}
\keyword{in}
\keyword{internal}
\keyword{into}
\keyword{is}
\keyword{it}
\keyword{its}
\keyword{line}
\keyword{list}
\keyword{name,}
\keyword{of}
\keyword{ones}
\keyword{return}
\keyword{seperated}
\keyword{session.}
\keyword{several}
\keyword{shiny}
\keyword{the}
\keyword{to}
\keyword{translated}
\keyword{used}
\keyword{values.}
\keyword{with}

