% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RProcessFinish.R
\name{RProcessFinishUI}
\alias{RProcessFinishUI}
\title{RProcessFinishUI}
\usage{
RProcessFinishUI(id)
}
\arguments{
\item{id}{chr id of this object for shiny session}
}
\value{
chr HTML for creating ui elements.
}
\description{
This is the UI part of a module to observe R scripts which are running as background processes from the shiny session.
It consists of only a red text field which will be visible as an error message if something during script execution went wrong.
}
\details{
For more details see the documentation of the corresponding \code{\link{RProcessFinish}} function.
}
\examples{

}
\seealso{
Seome examples on how to use RProcess module functions are given in the vignette \emph{RProcess Module Functions}
         (\code{vignette("RProcess_functions", package = "shinyTools")})

In the R script communications functions such as \code{\link{Rscript_Init}} and \code{\link{Rscript_Fin}} should be used
         to ensure correct communication between the R script and the shiny session.

Other RProcess module functions: \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{RProcessUI}},
  \code{\link{RProcess}}
}

