% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{RProcessInit}
\alias{RProcessInit}
\title{RProcessInit}
\usage{
RProcessInit(args = commandArgs(TRUE), log = paste("Starting at",
  Sys.time()))
}
\arguments{
\item{args}{chr of path to *.status file (commandArgs(TRUE)).
If the R script was started using RProcess module, this is the first trailing argument of Rscript (default).}

\item{log}{chr arr used as initial log entry for this process. Each element is a line.}
}
\value{
list of class ProcessInfo.
        This list contains information about the current process and is used by other functions of this module.
}
\description{
If a R script was started by shiny using the RProcess module, this should be the first command of the script.
This function reads the *.status file and returns it as a list of class ProcessInfo.
}
\details{
A *.status file is used for communication between the shiny session and the R batch process (started with Rscript).
Information provided by RProcess module is handed over to the R batch process via this file.
Conversely the R batch process writes on this file to hand over status updates to the shiny session.
By using the functions \code{\link{RProcessInit}}, \code{\link{RProcessUp}}, \code{\link{Log}}, \code{\link{Try}},
\code{\link{RProcessFinish}} it is ensured that the communication bewteen shiny session and R batch script work and
that no files are falsly overwritten.

If a log file exists, it is appended by the lines defined with log. Each element of log is a line.
Information about path of the log file, the session id and the process id is taken from the ProcessInfo object.
If a session id exists, every line is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the *.status file.
If they do not match, quit() is run.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Log}},
  \code{\link{RProcessUp}}, \code{\link{ReadInfo}},
  \code{\link{Try}}
}

