% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RProcess.R
\name{RProcessUI}
\alias{RProcessUI}
\title{RProcessUI}
\usage{
RProcessUI(id)
}
\arguments{
\item{id}{\code{chr} id of this object for shiny session}
}
\value{
chr HTML for creating ui elements.
}
\description{
UI part of a module which starts and observes a custom R script as background process.
The UI of this module is a red text field which pops up as error message if something while starting the script or during the script went wrong.
}
\details{
This module starts and then observes a custom R script.
 The communication between the shiny session and the R script is done via a \code{*.status} file.
 Input and output are handed via a \code{*.rds} file.

To ensure that the communication between the shiny session and the R script is working properly, use \code{\link{Rscript_Init}}
 to start the script and \code{\link{Rscript_Fin}} to finish it.
}
\seealso{
For further information see the documentation of Rscript communication such as \code{\link{Rscript_Init}}.
         Examples can be found in the vignette \emph{RProcess Module Functions}.

Other RProcess module functions: \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{RProcess}}
}

