% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Rscript_Fin}
\alias{Rscript_Fin}
\title{Rscript_Fin}
\usage{
Rscript_Fin(output, log = paste("Process finishes at", Sys.time()),
  info = NULL)
}
\arguments{
\item{output}{object which will be handed over back to shiny session}

\item{log}{\code{chr} arr used as log entry for this process (\code{paste("Process finishes at", Sys.time())}). Each element is a line.}

\item{info}{\emph{ProcessInfo} object or \code{NULL} (\code{NULL}).
This is a list of relevant information about the process.
It can be created with \code{\link{Rscript_Init}}.
If \code{NULL} the global environment is searched for an object of class \emph{ProcessInfo}.}
}
\description{
This function is used finish a R script run by a RProcess module such as \code{\link{RProcess}}.
It should be the last command of the script since it writes an output file, tells the shiny session about it,
and ends the R session with a call to \code{quit}.
}
\details{
The output will overwrite the input file used by the same R script.
By setting progress to 1 in the \code{*.status} file, the shiny session knows that this R script has finished.
It will then read the output file.

For correct communication with the shiny session an \emph{ProcessInfo} object given (argument \code{info}).
If a log file exists, it is appended with the message given in argument \code{log}. Each element is a line.
If a session id exists the log is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the \code{*.status} file.
If they do not match the R session ends with a call to \code{quit}.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\seealso{
These Rscript communication functions are used in R scripts started by a RProcess module such as \code{\link{RProcess}}.
         For some examples on how they work together see the vignette \emph{RProcess Module Functions}.

Other Rscript communication functions: \code{\link{ReadInfo}},
  \code{\link{Rscript_Abort}}, \code{\link{Rscript_Init}},
  \code{\link{Rscript_Log}}, \code{\link{Rscript_Start}},
  \code{\link{Rscript_Try}}, \code{\link{Rscript_Up}}
}

