% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Rscript_Log}
\alias{Rscript_Log}
\title{Rscript_Log}
\usage{
Rscript_Log(msg, info = NULL)
}
\arguments{
\item{msg}{chr arr with log messages. Each element is a line.}

\item{info}{ProcessInfo object or NULL (NULL).
This is a list of relevant information about the process.
It can be created with \code{\link{RProcessInit}}.
If NULL the global environment is searched for an object of class ProcessInfo.}
}
\value{
TRUE if a log was written, FALSE if not
}
\description{
This function is used to write log messages for a R script run by a RProcess module such as \code{\link{RProcess}}.
}
\details{
If a log file exists, it is appended by the lines defined in with msg. Each element of msg is a line.
Information about path of the log file, the session id and the process id is taken from the \code{ProcessInfo} object (see arguments).
If a session id exists, every line is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the \code{*.status} file.
If they do not match, the R session ends with a call to \code{quit}.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\seealso{
These Rscript communication functions are used in R scripts started by a RProcess module such as \code{\link{RProcess}}.
         For some examples on how they work together see the vignette \emph{RProcess Module Functions}.

Other Rscript communication functions: \code{\link{ReadInfo}},
  \code{\link{Rscript_Abort}}, \code{\link{Rscript_Fin}},
  \code{\link{Rscript_Init}}, \code{\link{Rscript_Start}},
  \code{\link{Rscript_Try}}, \code{\link{Rscript_Up}}
}

