% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Rscript_Start}
\alias{Rscript_Start}
\title{Rscript_Start}
\usage{
Rscript_Start(id, script, log = c(paste0("Start process with id ", id, ":"),
  script), logFile = NULL, info = NULL)
}
\arguments{
\item{id}{chr of process id. This id will be used by RProcessFinish to watch results of this process}

\item{script}{chr full path of R batch script which is to be started}

\item{log}{chr arr used as log entry in case logFile is not NULL. Each element is a line.}

\item{logFile}{chr or NULL (NULL) full path of logFile is needed. This log file will be used by the process started}

\item{info}{ProcessInfo object or NULL (NULL).
This is a list of relevant information about the process.
It can be created with \code{\link{RProcessInit}}.
If NULL the global environment is searched for an object of class ProcessInfo.}
}
\value{
TRUE
}
\description{
This function is used to write log messages for a R script run by a RProcess module such as \code{\link{RProcess}}.
}
\details{
This is a convenience function for a R batch script started by a shiny session with the RProcess module.
It should be the last command of a script since it writes an output file and send a termination singal.

The output will overwrite the input file used by the same R script.
By setting the progress to 1, the shiny session knows that this R script has finished.
It will then read the output file. The script quits itself.

The communication happens via a *.status file.
Information provided by RProcess module is handed over to the R batch process via this file.
Conversely the R batch process writes on this file to hand over status updates to the shiny session.
By using the functions \code{\link{Rscript_Init}}, \code{\link{Rscript_Up}}, \code{\link{Rscript_Log}}, \code{\link{Rscript_Try}},
\code{\link{Rscript_Fin}} it is ensured that the communication bewteen shiny session and R batch script work and
that no files are falsly overwritten.

If a log file exists, it is appended by the lines defined with log. Each element of log is a line.
Information about path of the log file, the session id and the process id is taken from the ProcessInfo object.
If a session id exists, every line is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the *.status file.
If they do not match, quit() is run.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\examples{

}
\seealso{
These Rscript communication functions are used in R scripts started by a RProcess module such as \code{\link{RProcess}}.
         For some examples on how they work together see the vignette \emph{RProcess Module Functions}.

Other Rscript communication functions: \code{\link{ReadInfo}},
  \code{\link{Rscript_Abort}}, \code{\link{Rscript_Fin}},
  \code{\link{Rscript_Init}}, \code{\link{Rscript_Log}},
  \code{\link{Rscript_Try}}, \code{\link{Rscript_Up}}
}

