% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Rscript_Try}
\alias{Rscript_Try}
\title{Rscript_Try}
\usage{
Rscript_Try(expr, ignore = FALSE, info = NULL)
}
\arguments{
\item{expr}{R expression to be evaluated}

\item{ignore}{\code{bool} (\code{FALSE}) whether the script should continue after an error occured}

\item{info}{\emph{ProcessInfo} object or \code{NULL} (\code{NULL}).
This is a list of relevant information about the process.
It can be created with \code{\link{Rscript_Init}}.
If NULL the global environment is searched for an object of class \emph{ProcessInfo}.}
}
\value{
\code{NULL} (if error occured) or result of evaluated expr
}
\description{
This function is used to recover an error in a R script run by a RProcess module such as \code{\link{RProcess}}.
In general, if the R script runs into an error its shiny session will never know about it.
This functions, reports this error to the shiny session and writes a log if desired.
}
\details{
For correct communication with the shiny session an \emph{ProcessInfo} object given (argument \code{info}).
If a log file exists, it is appended with the error message.
If a session id exists the log is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the \code{*.status} file.
If they do not match the R session ends with a call to \code{quit}.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\seealso{
These Rscript communication functions are used in R scripts started by a RProcess module such as \code{\link{RProcess}}.
         For some examples on how they work together see the vignette \emph{RProcess Module Functions}.

Other Rscript communication functions: \code{\link{ReadInfo}},
  \code{\link{Rscript_Abort}}, \code{\link{Rscript_Fin}},
  \code{\link{Rscript_Init}}, \code{\link{Rscript_Log}},
  \code{\link{Rscript_Start}}, \code{\link{Rscript_Up}}
}

