% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{Up}
\alias{Up}
\title{Up}
\usage{
Up(progress, log = paste("Progress update to", progress), info = NULL)
}
\arguments{
\item{progress}{num > 0 and < 1 indicating progress}

\item{log}{chr arr used as initial log entry for this process. Each element is a line.}

\item{info}{ProcessInfo object or NULL (NULL).
This is a list of relevant information about the process.
It can be created with \code{\link{RProcessInit}}.
If NULL the global environment is searched for an object of class ProcessInfo.}
}
\value{
list of class ProcessInfo.
        This list contains information about the current process and is used by other functions of this module.
}
\description{
This is a convenience function used in a R batch script for sending a status update to the shiny session which started it.
}
\details{
A progess update might be used by the RProcess module in the shiny session to provide a progress bar.
In addition a log entry about the progress update is written in case a log file was defined.
This information is taken from the ProcessInfo object.

The communication happens via a *.status file.
Information provided by RProcess module is handed over to the R batch process via this file.
Conversely the R batch process writes on this file to hand over status updates to the shiny session.
By using the functions \code{\link{Init}}, \code{\link{Up}}, \code{\link{Log}}, \code{\link{Try}},
\code{\link{Fin}} it is ensured that the communication bewteen shiny session and R batch script work and
that no files are falsly overwritten.

If a log file exists, it is appended by the lines defined with log. Each element of log is a line.
Information about path of the log file, the session id and the process id is taken from the ProcessInfo object.
If a session id exists, every line is prefixed with it.
If a process id exists, before writing anything the loaded process id is compared with the current one -- read from the *.status file.
If they do not match, quit() is run.
This ensures that, in case a newer version of the same process is running, the current one will quit and not overwrite anything.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Abort}},
  \code{\link{Fin}}, \code{\link{Init}}, \code{\link{Log}},
  \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{ReadInfo}},
  \code{\link{Start}}, \code{\link{Try}}
}

