% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessFunctions.R
\name{ReadInfo}
\alias{ReadInfo}
\title{ReadInfo}
\usage{
ReadInfo(statusFile)
}
\arguments{
\item{statusFile}{chr path to relevant *.status file}
}
\value{
list of class ProcessInfo.
        This list contains information about the current process and is used by other functions of this module.
}
\description{
Read content of a *.status file and return it as list of class ProcessInfo.
For creating the ProcessInfo object initially in a R script it is better to use \code{\link{RProcessInit}}.
It uses this function but also takes care of some other stuff.
}
\details{
The *.status file is used to communicate with the shiny session.
Each line in the file is translated into a list element with several entries seperated by ";".
The first entry is used as the element name, the following ones are its values.

By using the functions \code{\link{Init}}, \code{\link{Up}}, \code{\link{Log}}, \code{\link{Try}},
\code{\link{Fin}} it is ensured that the communication bewteen shiny session and R batch script work and
that no files are falsly overwritten.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Abort}},
  \code{\link{Fin}}, \code{\link{Init}}, \code{\link{Log}},
  \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{Start}},
  \code{\link{Try}}, \code{\link{Up}}
}

