% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RProcessFinish.R
\name{RProcessFinishUI}
\alias{RProcessFinishUI}
\title{RProcessFinishUI
UI part of a start module for R batch processes. The UI is only an error message which appears only if an error happened.}
\usage{
RProcessFinishUI(id)
}
\arguments{
\item{id}{chr id of this object for shiny session}
}
\value{
chr HTML for creating ui elements.
}
\description{
This is a process starter for R batch processes which are started from the shiny session, but should run in the background.
This is useful for things that take a while to compute. The app stays responsive while the batch process is running.
An R object can be passed through to a R batch script which will be started with 'Rscript'.
This object and information about the R batch script are saved as an *.rds and an *.status file.
These files are used for communication between both processes and they are written into the directory defined as 'pwd'.
The 2 files are prefixed with the module id.
}
\details{
This module starts and then observes a custom R script.
 The communication between the shiny session and the R script is done via a \code{*.status} file.
 Input and output are handed via a \code{*.rds} file.

To ensure that the communication between the shiny session and the R script is working properly, use \code{\link{Rscript_Init}}
 to start the script and \code{\link{Rscript_Fin}} to finish it.
}
\examples{

}
\seealso{
For further information see the documentation of Rscript communication such as \code{\link{Rscript_Init}}.
         Examples can be found in the vignette \emph{RProcess Module Functions}.

Other RProcess module functions: \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{RProcessUI}},
  \code{\link{RProcess}}
}

