% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RProcessStart.R
\name{RProcessStart}
\alias{RProcessStart}
\title{RProcessStart}
\usage{
RProcessStart(input, output, session, trigger, object, script, logFile = NULL,
  sessionid = NULL, pwd = getwd(), checkFun = NULL, addArgs = NULL)
}
\arguments{
\item{input}{argument used by shiny session}

\item{output}{argument used by shiny session}

\item{session}{argument used by shiny session}

\item{trigger}{reactive which starts the process (e.g. action button input)}

\item{object}{reactive object that will be handed over to R batch script as input}

\item{script}{chr path to R batch script (name of file included)}

\item{logFile}{chr or NULL (NULL) path of (desired) logging file (name included)}

\item{sessionid}{chr or NULL (NULL) id for the session. Log entries will be prefixed with it.}

\item{pwd}{chr (getwd()) path of R batch process working directory.
That's where intermediate files will appear as well.}

\item{checkFun}{chr or NULL (NULL) if not NULL name of a function which can be used as a quality check for object
right before it is handed ober to the R batch script}

\item{addArgs}{list or NULL (NULL) if not NULL list of additional arguments which will be passed to checkFun}
}
\description{
This is a process starter for R batch processes which are started from the shiny session, but should run in the background.
This is useful for things that take a while to compute. The app stays responsive while the batch process is running.
}
\details{
An R object can be passed through to a R batch script which will be started with 'Rscript'.
This object and information about the R batch script are saved as an *.rds and an *.status file.
These files are used for communication between both processes and they are written into the directory defined as 'pwd'.
The process is started with a reactive 'trigger'.

With argument checkFun a function name can be defined which will be used as quality control for 'object'.
This function is run before the R batch script is started.
First argument of this function is the 'object' handed over to RProcessStart.
The function must either return NULL or a chr value. NULL means the input is valid.
Thereby the module will start the R batch script.
If the input should not be valid, the function must return a character value.
This chr value will be rendered as a error message for the user, and the modul will not start the R batch script.

Additional argumets can be handed over to checkFun via the list addArgs.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Abort}},
  \code{\link{Fin}}, \code{\link{Init}}, \code{\link{Log}},
  \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}}, \code{\link{ReadInfo}},
  \code{\link{Start}}, \code{\link{Try}}, \code{\link{Up}}
}

