% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RProcessStart.R
\name{RProcessStartUI}
\alias{RProcessStartUI}
\title{RProcessStartUI}
\usage{
RProcessStartUI(id)
}
\arguments{
\item{id}{chr id of this object for shiny session}
}
\value{
chr HTML for creating ui elements.
}
\description{
UI part of a start module for R batch processes. The UI is only an error message which appears only if an error happened.
}
\details{
This is a process starter for R batch processes which are started from the shiny session, but should run in the background.
This is useful for things that take a while to compute. The app stays responsive while the batch process is running.
An R object can be passed through to a R batch script which will be started with 'Rscript'.
This object and information about the R batch script are saved as an *.rds and an *.status file.
These files are used for communication between both processes and they are written into the directory defined as 'pwd'.
The 2 files are prefixed with the module id.
}
\examples{

}
\seealso{
Other RProcess module functions: \code{\link{Abort}},
  \code{\link{Fin}}, \code{\link{Init}}, \code{\link{Log}},
  \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStart}}, \code{\link{ReadInfo}},
  \code{\link{Start}}, \code{\link{Try}}, \code{\link{Up}}
}

