% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RProcess.R
\name{RProcessUI}
\alias{RProcessUI}
\title{RProcessUI}
\usage{
RProcessUI(id)
}
\arguments{
\item{id}{\code{chr} id of this object for shiny session}
}
\value{
\code{chr} HTML for creating ui elements.
}
\description{
UI part of a module which starts and observes a custom R script as background process.
The UI of this module is a red text field which pops up as error message if something went wrong when starting the script or during script execution.
}
\details{
For more details see the documentation of the corresponding \code{\link{RProcess}} function.
}
\seealso{
Seome examples on how to use RProcess module functions are given in the vignette \emph{RProcess Module Functions}
         (\code{vignette("RProcess_functions", package = "shinyTools")})

In the R script communications functions such as \code{\link{Rscript_Init}} and \code{\link{Rscript_Fin}} should be used
         to ensure correct communication between the R script and the shiny session.

Other RProcess module functions: \code{\link{RProcessFinishUI}},
  \code{\link{RProcessFinish}},
  \code{\link{RProcessStartUI}},
  \code{\link{RProcessStart}}, \code{\link{RProcess}}
}

